Bridge previous dragonpilot integration onto commaai/openpilot master
Preserve the completed dragonpilot tree at 0858ec6b080fa09c754cfed7516f044e0613493b while attaching it to the real upstream commit 08fd571272513ce792b4882857b21b1731f1252b for future incremental merges.

+++ tools/release/build_release.sh
@@ -18,17 +18,24 @@ fi
 
 BUILD_BRANCH=release-mici-staging
 
+if [ -z "$SOURCE_BRANCH" ]; then
+  echo "SOURCE_BRANCH is not set"
+  exit 1
+fi
+
 
 # set git identity
-source $DIR/identity.sh
+source /data/identity.sh
 
 echo "[-] Setting up repo T=$SECONDS"
 rm -rf $BUILD_DIR
 mkdir -p $BUILD_DIR
 cd $BUILD_DIR
 git init
-git remote add origin git@github.com:commaai/openpilot.git
-git checkout --orphan $BUILD_BRANCH
+# git remote add origin git@github.com:commaai/openpilot.git
+# git checkout --orphan $BUILD_BRANCH
+git remote add origin https://github.com/dragonpilot/dev.git
+git checkout --orphan $SOURCE_BRANCH
 
 # do the files copy
 echo "[-] copying files T=$SECONDS"
@@ -44,19 +51,25 @@ rm -f panda/board/obj/panda_h7.bin.signed
 VERSION=$(cat openpilot/common/version.h | awk -F[\"-]  '{print $2}')
 echo "[-] committing version $VERSION T=$SECONDS"
 git add -f .
-git commit -a -m "openpilot v$VERSION release"
+git commit -a -m "dragonpilot v$VERSION release"
 
 # Build
 export PYTHONPATH="$BUILD_DIR"
 scons
 
-if [ -z "$PANDA_DEBUG_BUILD" ]; then
-  # release panda fw
-  CERT=/data/pandaextra/certs/release RELEASE=1 scons panda/
-else
-  # build with ALLOW_DEBUG=1 to enable features like experimental longitudinal
-  scons panda/
-fi
+# if [ -z "$PANDA_DEBUG_BUILD" ]; then
+#   # release panda fw
+#   CERT=/data/pandaextra/certs/release RELEASE=1 scons panda/
+# else
+#   # build with ALLOW_DEBUG=1 to enable features like experimental longitudinal
+#   scons panda/
+# fi
+scons -j$(nproc) panda/
+
+# panda tici
+rm -f panda_tici/board/obj/panda.bin.signed
+rm -f panda_tici/board/obj/panda_h7.bin.signed
+scons -j$(nproc) panda_tici/
 
 # Ensure no submodules in release
 if test "$(git submodule--helper list | wc -l)" -gt "0"; then
@@ -79,20 +92,34 @@ rm -f openpilot/selfdrive/modeld/models/*.onnx*
 # Mark as prebuilt release
 touch prebuilt
 
+# dragonpilot customized
+find . -name '*.cc' -delete
+find openpilot/selfdrive/ui/ -name '*.h' -delete
+# rick - some test codes are used in the code
+# find . -type d -name "tests" -exec rm -rf {} +
+find . -type d -name 'x86_64' -exec rm -rf {} +
+find . -type d -name 'Darwin' -exec rm -rf {} +
+rm -fr tinygrad_repo/docs/tinygrad_intro.pdf # 1.9M
+rm -fr openpilot/cereal/gen/cpp/log.capnp.h # 2.5M
+rm -fr tinygrad_repo/extra/hip_gpu_driver/gc_10_3_0_offset.h # 1.4 M
+rm -fr tinygrad_repo/extra/accel/tpu/logs/tpu_driver.t1v-n-852cd0d5-w-0.taylor.log.INFO.20210619-062914.26926.gz # 1.3 M
+
 # Add built files to git
 git add -f .
-git commit --amend -m "openpilot v$VERSION"
+git commit --amend -m "dragonpilot v$VERSION"
+
+git branch -m $RELEASE_BRANCH
 
 # Run tests
-cd $BUILD_DIR
-RELEASE=1 ./openpilot/selfdrive/test/test_onroad.py
-#pytest openpilot/selfdrive/car/tests/test_car_interfaces.py
+# cd $BUILD_DIR
+# RELEASE=1 ./openpilot/selfdrive/test/test_onroad.py
+# pytest openpilot/selfdrive/car/tests/test_car_interfaces.py
 
-echo "[-] pushing release T=$SECONDS"
-REFS=()
-for branch in ${RELEASE_BRANCH//,/ }; do
-  REFS+=("$BUILD_BRANCH:$branch")
-done
-git push -f origin "${REFS[@]}"
+# echo "[-] pushing release T=$SECONDS"
+# REFS=()
+# for branch in ${RELEASE_BRANCH//,/ }; do
+#   REFS+=("$BUILD_BRANCH:$branch")
+# done
+# git push -f origin "${REFS[@]}"
 
 echo "[-] done T=$SECONDS"
